+++ src/decorators/OnDisconnect.ts
@@ -1,11 +1,11 @@
 import { addActionToControllerMetadata } from '../util/add-action-to-controller-metadata';
-import { ActionType } from '../types/enums/ActionType';
+import { SocketEventType } from '../types/enums/SocketEventType';
 
 export function OnDisconnect(): Function {
   return function (object: Object, methodName: string) {
     addActionToControllerMetadata(object.constructor, {
       methodName,
-      type: ActionType.DISCONNECT,
+      type: SocketEventType.DISCONNECT,
       options: {},
     });
   };