+++ src/types/enums/SocketEventType.ts
@@ -1,4 +1,4 @@
-export enum ActionType {
+export enum SocketEventType {
   MESSAGE,
   CONNECT,
   DISCONNECT,